Move nodes from logic.rs to text/src/lib.rs
