+++ node-graph/nodes/gcore/Cargo.toml
@@ -27,11 +27,6 @@ node-macro = { workspace = true }
 dyn-any = { workspace = true }
 glam = { workspace = true }
 log = { workspace = true }
-serde_json = { workspace = true }
-convert_case = { workspace = true }
-titlecase = { workspace = true }
-fancy-regex = { workspace = true }
-unicode-segmentation = { workspace = true }
 
 # Optional workspace dependencies
 serde = { workspace = true, optional = true }
