+++ node-graph/nodes/math/src/lib.rs
@@ -1,5 +1,5 @@
 use core_types::Context;
-use core_types::registry::types::{Fraction, Percentage, PixelSize, TextArea};
+use core_types::registry::types::{Fraction, Percentage, PixelSize};
 use core_types::table::Table;
 use core_types::transform::Footprint;
 use core_types::{Color, Ctx, num_traits};
@@ -871,12 +871,6 @@ fn sample_gradient(_: impl Ctx, _primary: (), gradient: Table<GradientStops>, po
 	let position = position.clamp(0., 1.);
 	let color = row.element.evaluate(position);
 	Table::new_from_element(color)
-}
-
-/// Constructs a string value which may be set to any plain text.
-#[node_macro::node(category("Value"))]
-fn string_value(_: impl Ctx, _primary: (), string: TextArea) -> String {
-	string
 }
 
 /// Constructs a footprint value which may be set to any transformation of a unit square describing a render area, and a render resolution at least 1x1 integer pixels.
